[backport] net: change CNetAddr::ip to have flexible size
Summary
---

This is a backport of https://github.com/bitcoin/bitcoin/pull/19628/commits/102867c587f5f7954232fb8ed8e85cda78bb4d32

> Before this change `CNetAddr::ip` was a fixed-size array of 16 bytes,
> not being able to store larger addresses (e.g. TORv3) and encoded
> smaller ones as 16-byte IPv6 addresses.
>
> Change its type to `prevector`, so that it can hold larger addresses
> and do not disguise non-IPv6 addresses as IPv6. So the IPv4 address
> `1.2.3.4` is now encoded as `01020304` instead of
> `00000000000000000000FFFF01020304`.
>
> Rename `CNetAddr::ip` to `CNetAddr::m_addr` because it is not an "IP"
> or "IP address" (TOR addresses are not IP addresses).
>
> In order to preserve backward compatibility with serialization (where
> e.g. `1.2.3.4` is serialized as `00000000000000000000FFFF01020304`)
> introduce `CNetAddr` dedicated legacy serialize/unserialize methods.
>
> Adjust `CSubNet` accordingly. Still use `CSubNet::netmask[]` of fixed
> 16 bytes, but use the first 4 for IPv4 (not the last 4). Only allow
> subnetting for IPv4 and IPv6.
>

Test plan
---

* `ninja all check-all`

+++ src/netbase.cpp
@@ -17,6 +17,8 @@
 #include <tinyformat.h>
 
 #include <atomic>
+#include <cstdint>
+#include <limits>
 
 #ifndef WIN32
 #include <fcntl.h>
@@ -901,8 +903,8 @@ bool LookupSubNet(const std::string &strSubnet, CSubNet &ret) {
         CNetAddr network = vIP[0];
         if (slash != strSubnet.npos) {
             std::string strNetmask = strSubnet.substr(slash + 1);
-            int32_t n;
-            if (ParseInt32(strNetmask, &n)) {
+            uint8_t n;
+            if (ParseUInt8(strNetmask, &n)) {
                 // If valid number, assume CIDR variable-length subnet masking
                 ret = CSubNet(network, n);
                 return ret.IsValid();
